feat(metrics): 将心率和信号质量指标改为浮点数精度
- 将 MetricsData 结构体中的 heartRate 和 sqi 字段从 int 类型改为 double 类型
- 更新 MetricsData 构造函数以支持浮点数值初始化
- 修改 MetricsPanel 中的数据显示逻辑，添加小数位截取和格式化功能
- 将标题文本从"实时监测指标"改为"实时信号"
- 更新 MetricsService 中的数据解析逻辑，支持 JSON 中的浮点数值
- 修改模拟数据生成功能，使用随机浮点数代替整数
- 更新文档注释，说明期望的 JSON 格式包含浮点数值

+++ src/model/MetricsData.h
@@ -8,14 +8,14 @@
  * 在 Service、Controller 与 View 之间传递，不持有状态，不触发副作用。
  */
 struct MetricsData {
-    int heartRate;  ///< 心率，单位 bpm
-    int sqi;        ///< 信号质量指数，0-100
+    double heartRate;  ///< 心率，单位 bpm
+    double sqi;        ///< 信号质量指数，0-100
 
     MetricsData()
-        : heartRate(0), sqi(0)
+        : heartRate(0.0), sqi(0.0)
     {}
 
-    MetricsData(int hr, int sq)
+    MetricsData(double hr, double sq)
         : heartRate(hr), sqi(sq)
     {}
 };
